Split seat and sector components from SenateTab
Split out the SenateSeat and SenateSector components from SenateTab file because I prefer to have one component per file.
